Data.Array.Accelerate.IO.Codec.Picture.Types

Description

Documentation

data Image a

The main type of this package, one that most functions work on, is Image.

Parameterized by the underlying pixel format it forms a rigid type. If you wish to store images of different or unknown pixel formats use DynamicImage.

Image is essentially a rectangular pixel buffer of specified width and height. The coordinates are assumed to start from the upper-left corner of the image, with the horizontal position first and vertical second.

Constructors

Image

Fields

imageWidth :: !Int

Width of the image in pixels

imageHeight :: !Int

Height of the image in pixels.

imageData :: Vector (PixelBaseComponent a)

Image pixel data. To extract pixels at a given position you should use the helper functions.

Internally pixel data is stored as consecutively packed lines from top to bottom, scanned from left to right within individual lines, from first to last color component within each pixel.

class (Storable (PixelBaseComponent a), Num (PixelBaseComponent a), Eq a) => Pixel a

Definition of pixels used in images. Each pixel has a color space, and a representative component (Word8 or Float).

Minimal complete definition

mixWith, pixelOpacity, componentCount, colorMap, pixelAt, readPixel, writePixel, unsafePixelAt, unsafeReadPixel, unsafeWritePixel

Associated Types

type PixelBaseComponent a :: Type

Type of the pixel component, "classical" images would have Word8 type as their PixelBaseComponent, HDR image would have Float for instance

type Pixel8 = Word8

Type alias for 8bit greyscale pixels. For simplicity, greyscale pixels use plain numbers instead of a separate type.

type Pixel16 = Word16

Type alias for 16bit greyscale pixels.

type Pixel32 = Word32

Type alias for 32bit greyscale pixels.

type PixelF = Float

Type alias for 32bit floating point greyscale pixels. The standard bounded value range is mapped to the closed interval [0,1] i.e.

map promotePixel [0, 1 .. 255 :: Pixel8] == [0/255, 1/255 .. 1.0 :: PixelF]

data PixelYA8

Pixel type storing 8bit Luminance (Y) and alpha (A) information. Values are stored in the following order:

Luminance

Alpha

Constructors

PixelYA8 !Pixel8 !Pixel8

pattern PixelYA8_ :: Exp Pixel8 -> Exp Pixel8 -> Exp PixelYA8

data PixelYA16

Pixel type storing 16bit Luminance (Y) and alpha (A) information. Values are stored in the following order:

Luminance

Alpha

Constructors

PixelYA16 !Pixel16 !Pixel16

pattern PixelYA16_ :: Exp Pixel16 -> Exp Pixel16 -> Exp PixelYA16

data PixelRGB8

Classic pixel type storing 8bit red, green and blue (RGB) information. Values are stored in the following order:

Red

Green

Blue

Constructors

PixelRGB8 !Pixel8 !Pixel8 !Pixel8

pattern PixelRGB8_ :: Exp Pixel8 -> Exp Pixel8 -> Exp Pixel8 -> Exp PixelRGB8

data PixelRGB16

Pixel type storing 16bit red, green and blue (RGB) information. Values are stored in the following order:

Red

Green

Blue

Constructors

PixelRGB16 !Pixel16 !Pixel16 !Pixel16

pattern PixelRGB16_ :: Exp Pixel16 -> Exp Pixel16 -> Exp Pixel16 -> Exp PixelRGB16

data PixelRGBF

HDR pixel type storing floating point 32bit red, green and blue (RGB) information. Same value range and comments apply as for PixelF. Values are stored in the following order:

Red

Green

Blue

Constructors

PixelRGBF !PixelF !PixelF !PixelF

pattern PixelRGBF_ :: Exp PixelF -> Exp PixelF -> Exp PixelF -> Exp PixelRGBF

data PixelRGBA8

Classical pixel type storing 8bit red, green, blue and alpha (RGBA) information. Values are stored in the following order:

Red

Green

Blue

Alpha

Constructors

PixelRGBA8 !Pixel8 !Pixel8 !Pixel8 !Pixel8

pattern PixelRGBA8_ :: Exp Pixel8 -> Exp Pixel8 -> Exp Pixel8 -> Exp Pixel8 -> Exp PixelRGBA8

data PixelRGBA16

Pixel type storing 16bit red, green, blue and alpha (RGBA) information. Values are stored in the following order:

Red

Green

Blue

Alpha

Constructors

PixelRGBA16 !Pixel16 !Pixel16 !Pixel16 !Pixel16

pattern PixelRGBA16_ :: Exp Pixel16 -> Exp Pixel16 -> Exp Pixel16 -> Exp Pixel16 -> Exp PixelRGBA16

data PixelCMYK8

Pixel type storing 8bit cyan, magenta, yellow and black (CMYK) information. Values are stored in the following order:

Cyan

Magenta

Yellow

Black

Constructors

PixelCMYK8 !Pixel8 !Pixel8 !Pixel8 !Pixel8

pattern PixelCMYK8_ :: Exp Pixel8 -> Exp Pixel8 -> Exp Pixel8 -> Exp Pixel8 -> Exp PixelCMYK8

data PixelCMYK16

Pixel type storing 16bit cyan, magenta, yellow and black (CMYK) information. Values are stored in the following order:

Cyan

Magenta

Yellow

Black

Constructors

PixelCMYK16 !Pixel16 !Pixel16 !Pixel16 !Pixel16

pattern PixelCMYK16_ :: Exp Pixel16 -> Exp Pixel16 -> Exp Pixel16 -> Exp Pixel16 -> Exp PixelCMYK16

data PixelYCbCr8

Pixel type storing 8bit luminance, blue difference and red difference (YCbCr) information. Values are stored in the following order:

Y (luminance)

Cb

Cr

Constructors

PixelYCbCr8 !Pixel8 !Pixel8 !Pixel8

pattern PixelYCbCr8_ :: Exp Pixel8 -> Exp Pixel8 -> Exp Pixel8 -> Exp PixelYCbCr8

data PixelYCbCrK8

Pixel type storing value for the YCCK color space:

Y (Luminance)

Cb

Cr

Black

Constructors

PixelYCbCrK8 !Pixel8 !Pixel8 !Pixel8 !Pixel8

pattern PixelYCbCrK8_ :: Exp Pixel8 -> Exp Pixel8 -> Exp Pixel8 -> Exp Pixel8 -> Exp PixelYCbCrK8

Orphan instances

Elt PixelYA8

Instance details

Associated Types

type EltRepr PixelYA8 :: Type

Methods

eltType :: TupleType (EltRepr PixelYA8)

fromElt :: PixelYA8 -> EltRepr PixelYA8

toElt :: EltRepr PixelYA8 -> PixelYA8

Elt PixelYA16

Instance details

Associated Types

type EltRepr PixelYA16 :: Type

Methods

eltType :: TupleType (EltRepr PixelYA16)

fromElt :: PixelYA16 -> EltRepr PixelYA16

toElt :: EltRepr PixelYA16 -> PixelYA16

Elt PixelRGB8

Instance details

Associated Types

type EltRepr PixelRGB8 :: Type

Methods

eltType :: TupleType (EltRepr PixelRGB8)

fromElt :: PixelRGB8 -> EltRepr PixelRGB8

toElt :: EltRepr PixelRGB8 -> PixelRGB8

Elt PixelYCbCrK8

Instance details

Associated Types

type EltRepr PixelYCbCrK8 :: Type

Methods

eltType :: TupleType (EltRepr PixelYCbCrK8)

fromElt :: PixelYCbCrK8 -> EltRepr PixelYCbCrK8

toElt :: EltRepr PixelYCbCrK8 -> PixelYCbCrK8

Elt PixelRGB16

Instance details

Associated Types

type EltRepr PixelRGB16 :: Type

Methods

eltType :: TupleType (EltRepr PixelRGB16)

fromElt :: PixelRGB16 -> EltRepr PixelRGB16

toElt :: EltRepr PixelRGB16 -> PixelRGB16

Elt PixelRGBF

Instance details

Associated Types

type EltRepr PixelRGBF :: Type

Methods

eltType :: TupleType (EltRepr PixelRGBF)

fromElt :: PixelRGBF -> EltRepr PixelRGBF

toElt :: EltRepr PixelRGBF -> PixelRGBF

Elt PixelYCbCr8

Instance details

Associated Types

type EltRepr PixelYCbCr8 :: Type

Methods

eltType :: TupleType (EltRepr PixelYCbCr8)

fromElt :: PixelYCbCr8 -> EltRepr PixelYCbCr8

toElt :: EltRepr PixelYCbCr8 -> PixelYCbCr8

Elt PixelCMYK8

Instance details

Associated Types

type EltRepr PixelCMYK8 :: Type

Methods

eltType :: TupleType (EltRepr PixelCMYK8)

fromElt :: PixelCMYK8 -> EltRepr PixelCMYK8

toElt :: EltRepr PixelCMYK8 -> PixelCMYK8

Elt PixelCMYK16

Instance details

Associated Types

type EltRepr PixelCMYK16 :: Type

Methods

eltType :: TupleType (EltRepr PixelCMYK16)

fromElt :: PixelCMYK16 -> EltRepr PixelCMYK16

toElt :: EltRepr PixelCMYK16 -> PixelCMYK16

Elt PixelRGBA8

Instance details

Associated Types

type EltRepr PixelRGBA8 :: Type

Methods

eltType :: TupleType (EltRepr PixelRGBA8)

fromElt :: PixelRGBA8 -> EltRepr PixelRGBA8

toElt :: EltRepr PixelRGBA8 -> PixelRGBA8

Elt PixelRGBA16

Instance details

Associated Types

type EltRepr PixelRGBA16 :: Type

Methods

eltType :: TupleType (EltRepr PixelRGBA16)

fromElt :: PixelRGBA16 -> EltRepr PixelRGBA16

toElt :: EltRepr PixelRGBA16 -> PixelRGBA16

IsProduct Elt PixelYA8

Instance details

Associated Types

type ProdRepr PixelYA8 :: Type

Methods

fromProd :: PixelYA8 -> ProdRepr PixelYA8

toProd :: ProdRepr PixelYA8 -> PixelYA8

prod :: ProdR Elt (ProdRepr PixelYA8)

IsProduct Elt PixelYA16

Instance details

Associated Types

type ProdRepr PixelYA16 :: Type

Methods

fromProd :: PixelYA16 -> ProdRepr PixelYA16

toProd :: ProdRepr PixelYA16 -> PixelYA16

prod :: ProdR Elt (ProdRepr PixelYA16)

IsProduct Elt PixelRGB8

Instance details

Associated Types

type ProdRepr PixelRGB8 :: Type

Methods

fromProd :: PixelRGB8 -> ProdRepr PixelRGB8

toProd :: ProdRepr PixelRGB8 -> PixelRGB8

prod :: ProdR Elt (ProdRepr PixelRGB8)

IsProduct Elt PixelYCbCrK8

Instance details

Associated Types

type ProdRepr PixelYCbCrK8 :: Type

Methods

fromProd :: PixelYCbCrK8 -> ProdRepr PixelYCbCrK8

toProd :: ProdRepr PixelYCbCrK8 -> PixelYCbCrK8

prod :: ProdR Elt (ProdRepr PixelYCbCrK8)

IsProduct Elt PixelRGB16

Instance details

Associated Types

type ProdRepr PixelRGB16 :: Type

Methods

fromProd :: PixelRGB16 -> ProdRepr PixelRGB16

toProd :: ProdRepr PixelRGB16 -> PixelRGB16

prod :: ProdR Elt (ProdRepr PixelRGB16)

IsProduct Elt PixelRGBF

Instance details

Associated Types

type ProdRepr PixelRGBF :: Type

Methods

fromProd :: PixelRGBF -> ProdRepr PixelRGBF

toProd :: ProdRepr PixelRGBF -> PixelRGBF

prod :: ProdR Elt (ProdRepr PixelRGBF)

IsProduct Elt PixelYCbCr8

Instance details

Associated Types

type ProdRepr PixelYCbCr8 :: Type

Methods

fromProd :: PixelYCbCr8 -> ProdRepr PixelYCbCr8

toProd :: ProdRepr PixelYCbCr8 -> PixelYCbCr8

prod :: ProdR Elt (ProdRepr PixelYCbCr8)

IsProduct Elt PixelCMYK8

Instance details

Associated Types

type ProdRepr PixelCMYK8 :: Type

Methods

fromProd :: PixelCMYK8 -> ProdRepr PixelCMYK8

toProd :: ProdRepr PixelCMYK8 -> PixelCMYK8

prod :: ProdR Elt (ProdRepr PixelCMYK8)

IsProduct Elt PixelCMYK16

Instance details

Associated Types

type ProdRepr PixelCMYK16 :: Type

Methods

fromProd :: PixelCMYK16 -> ProdRepr PixelCMYK16

toProd :: ProdRepr PixelCMYK16 -> PixelCMYK16

prod :: ProdR Elt (ProdRepr PixelCMYK16)

IsProduct Elt PixelRGBA8

Instance details

Associated Types

type ProdRepr PixelRGBA8 :: Type

Methods

fromProd :: PixelRGBA8 -> ProdRepr PixelRGBA8

toProd :: ProdRepr PixelRGBA8 -> PixelRGBA8

prod :: ProdR Elt (ProdRepr PixelRGBA8)

IsProduct Elt PixelRGBA16

Instance details

Associated Types

type ProdRepr PixelRGBA16 :: Type

Methods

fromProd :: PixelRGBA16 -> ProdRepr PixelRGBA16

toProd :: ProdRepr PixelRGBA16 -> PixelRGBA16

prod :: ProdR Elt (ProdRepr PixelRGBA16)